# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go DDD Hexagonal architecture skeleton project API",
        "title": "Skeleton API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List notifications with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "List notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by user ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (pending, queued, sending, sent, delivered, failed)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by channel (email, sms, push, in_app)",
                        "name": "channel",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notifications list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_notifications_ports_http.NotificationResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new notification manually (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Create a notification",
                "parameters": [
                    {
                        "description": "Notification data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_notifications_ports_http.CreateNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Notification created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/notifications/preferences": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get current user's notification preferences",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get user notification preferences",
                "responses": {
                    "200": {
                        "description": "User preferences",
                        "schema": {
                            "$ref": "#/definitions/internal_notifications_ports_http.NotificationPreferencesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update current user's notification preferences",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Update user notification preferences",
                "parameters": [
                    {
                        "description": "Preferences data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_notifications_ports_http.NotificationPreferencesRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Preferences updated"
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/notifications/templates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List all notification templates optionally filtered by channel (admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "List notification templates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by channel (email, sms, push, in_app)",
                        "name": "channel",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Templates list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_notifications_ports_http.TemplateResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new notification template (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Create notification template",
                "parameters": [
                    {
                        "description": "Template data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_notifications_ports_http.CreateTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Template created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/notifications/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get notification details by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get notification by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification details",
                        "schema": {
                            "$ref": "#/definitions/internal_notifications_ports_http.NotificationResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_basilex_skeleton_pkg_apierror.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_basilex_skeleton_pkg_apierror.APIError": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "internal_notifications_ports_http.NotificationResponse": {
            "type": "object",
            "properties": {
                "attempts": {
                    "type": "integer"
                },
                "channel": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "delivered_at": {
                    "type": "string"
                },
                "device_token": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "failed_at": {
                    "type": "string"
                },
                "html_content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "max_attempts": {
                    "type": "integer"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "phone": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "sent_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token (format: \"Bearer {token}\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SessionAuth": {
            "type": "apiKey",
            "name": "session",
            "in": "cookie"
        }
    },
...
}